Allow dind-cluster.sh and dind-cluster-v1.11.sh to run on the latest Docker for Mac
The latest versions of Docker for Mac have changed what directories by default are
setup for bind volume mounts. The dind-cluster*.sh scripts assume that /boot and
/lib/modules can be bind mounted, but this is not the case with the new Docker for
Mac versions.

This commit detects if the script is running on a Mac, and if so, it replaces /boot
and /lib/modules with /tmp, which is by default bind mountable on the latest Docker
for Mac.

Tested on Docker for Mac version 2.0.0.0-beta1-mac75 (build 27117) with Kubernetes
version 1.11.0.

Signed-off-by: Kyle Mestery <mestery@mestery.com>